Move ContactList and add tests
